backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
python-multipart
python-dotenv
pydantic
email-validator
pydantic-settings
